requirements.txt
Click==7.0
dlib==19.16.0
face-recognition==1.2.3
face-recognition-models==0.3.0
Flask==1.0.2
Flask-Script==2.0.6
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
numpy==1.15.4
Pillow==5.4.1
Werkzeug==0.14.1
